+++ change_log.md
@@ -5,7 +5,7 @@
 * (done) User clicks a manual income button
 * (done) The Manual income button can be upgraded
 * (done) The user can buy and upgrade auto clickers
-* () auto clicker upgrades should have an effect on time
+* () use the dim return function to help set the auto clicker time
 * () per hour rates should reflect the current state of the auto clicker